% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ch_dmr.R
\name{chDMR}
\alias{chDMR}
\title{chDMR}
\usage{
chDMR(dmw, p = 0.05, min_size = 60000, gap1_merge = 5000, gap2_merge = 1e+05)
}
\arguments{
\item{dmw}{A \code{GRanges} object output from the \code{chDMW} function.}

\item{p}{A \code{numeric} P-value threshold for selecting candidate differentially methylated windows. Default is 0.05.}

\item{min_size}{An \code{integer} of minimum CH-DMR size.}

\item{gap1_merge}{An \code{integer} representing the distance to merge deferentially methylated windows to create DMRs.}

\item{gap2_merge}{An \code{integer} representing the distance to merge candidate DMRs.}
}
\value{
An object of class \code{GRanges}
}
\description{
Find contiguous regions of significant changes with same direction
}
