% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ch_dmr.R
\name{chDMW}
\alias{chDMW}
\title{chDMR}
\usage{
chDMW(
  pData,
  p = 0.05,
  bsgenome = BSgenome.Hsapiens.UCSC.hg19,
  contigs = paste0("chr", c(1:22)),
  cores = 2L,
  normalize = TRUE,
  normalize_method = "quantile"
)
}
\arguments{
\item{pData}{a \code{data.frame} object with the columns "sample", "group" and "bigwig_path". The "sample" column contains the unique identifiers for each sample. The "group" column contains the identifiers used to define groups for differential testing. The "bigwig_path" column contains the absolute file path for bigwig files created using the \code{bsseq_to_window_bigwig} function.}

\item{normalize_method}{}
}
\value{

}
\description{
Function for detecting CH (non-CG) differentially methylated regions (DMRs)
}
