% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_functions.R
\name{gr_to_loci}
\alias{gr_to_loci}
\title{gr_to_loci}
\usage{
gr_to_loci(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
Character.
}
\description{
Convert a GRanges object to a character vector of loci. These are useful as row.names for data.frame objects, pasting into a genome browser, among other things.
}
\examples{
gr <- GRanges(seqnames=Rle(c('chr1'), c(3)), IRanges(1:3, width=5))
gr_to_loci(gr)

}
